Add partialView ListPdtRatioAndValue

+++ BPMS01UnitTests/Standard_product_value.cs
@@ -26,7 +26,7 @@ namespace BPMS01UnitTests
             //action
             decimal std_pdt_price = GetQuota.GetStdPdtValue(bridgeStructure_type, bridgeLength, bridgeWidth, bridgeNspan, inspection_type);
 
-            double key=35000+120*(35-30)+120*(42-40)+0*(2 - 1);    //答案
+            decimal key=35000+120*(35-30)+120*(42-40)+0*(2 - 1);    //答案
             //assert
             Assert.AreEqual(std_pdt_price, key);
         }
